ts_setup: renamed from ts_find + included ts_config inside ts_setup
