+++ doc/Makefile.am
@@ -1,1 +1,1 @@
-dist_man_MANS = ts.conf.5 ts_test.1 ts_test_mt.1 ts_calibrate.1 ts_uinput.1 ts_print.1 ts_print_raw.1 ts_print_mt.1 ts_print_raw_mt.1 ts_finddev.1 ts_harvest.1 ts_read.3 ts_read_mt.3 ts_read_raw.3 ts_read_raw_mt.3 ts_open.3 ts_close.3 ts_config.3 ts_find.3
+dist_man_MANS = ts.conf.5 ts_test.1 ts_test_mt.1 ts_calibrate.1 ts_uinput.1 ts_print.1 ts_print_raw.1 ts_print_mt.1 ts_print_raw_mt.1 ts_finddev.1 ts_harvest.1 ts_read.3 ts_read_mt.3 ts_read_raw.3 ts_read_raw_mt.3 ts_open.3 ts_close.3 ts_config.3 ts_setup.3
