+++ doc/ts_config.3
@@ -25,5 +25,6 @@ Zero is returned on success. A negative value is returned in case of an error.
 .SH SEE ALSO
 .BR ts_read (3),
 .BR ts_open (3),
+.BR ts_setup (3),
 .BR ts_close (3),
 .BR ts.conf (5)
