--- doc/ts_find.3
@@ -1,40 +1,0 @@
-.TH TS_FIND 3  "" "" "tslib"
-.SH NAME
-ts_find \- find and open a touch screen input device
-.SH SYNOPSIS
-.nf
-.B #include <tslib.h>
-.sp
-.BI "struct tsdev *ts_find(int " nonblock ");"
-.sp
-.fi
-
-.SH DESCRIPTION
-.BR ts_find ()
-will try find and open this touch screen device.
-On success it allocates memory for a new touch screen device, opens and links the given input event device to it.
-
-When \fBTSLIB_TSDEVICE\fR is defined
-.BR ts_find ()
-will try to use device defined by it, otherwise following devices are tested:
-.RS
-/dev/input/ts,
-.br
-/dev/input/touchscreen,
-.br
-/dev/input/event0,
-.br
-/dev/touchscreen/ucb1x00.
-.RE
-
-.SH RETURN VALUE
-A pointer to the opened
-.BI "struct tsdev"
-is returned.
-
-.SH SEE ALSO
-.BR ts_open (3),
-.BR ts_close (3),
-.BR ts_read (3),
-.BR ts_config (3),
-.BR ts.conf (5)
