+++ src/Makefile.am
@@ -17,7 +17,7 @@ include_HEADERS  = tslib.h
 lib_LTLIBRARIES  = libts.la
 libts_la_SOURCES = ts_attach.c ts_close.c ts_config.c ts_error.c \
 		   ts_fd.c ts_load_module.c ts_open.c ts_parse_vars.c \
-		   ts_read.c ts_read_raw.c ts_option.c ts_find.c $(srcdir)/../plugins/plugins.h
+		   ts_read.c ts_read_raw.c ts_option.c ts_setup.c $(srcdir)/../plugins/plugins.h
 
 if ENABLE_STATIC_LINEAR_MODULE
 libts_la_SOURCES += $(top_srcdir)/plugins/linear.c
