+++ tests/ts_harvest.c
@@ -67,15 +67,10 @@ int main()
 	signal(SIGINT, sig);
 	signal(SIGTERM, sig);
 
-	ts = ts_find(0);
+	ts = ts_setup(NULL, 0);
 
 	if (!ts) {
-		perror("ts_open");
-		exit(1);
-	}
-
-	if (ts_config(ts)) {
-		perror("ts_config");
+		perror("ts_setup");
 		exit(1);
 	}
