+++ tests/ts_print.c
@@ -18,22 +18,10 @@
 int main()
 {
 	struct tsdev *ts;
-	char *tsdevice=NULL;
 
-	if( (tsdevice = getenv("TSLIB_TSDEVICE")) != NULL ) {
-		ts = ts_open(tsdevice,0);
-	} else {
-		if (!(ts = ts_open("/dev/input/event0", 0)))
-			ts = ts_open("/dev/touchscreen/ucb1x00", 0);
-	}
-
+	ts = ts_setup(NULL, 0);
 	if (!ts) {
-		perror("ts_open");
-		exit(1);
-	}
-
-	if (ts_config(ts)) {
-		perror("ts_config");
+		perror("ts_setup");
 		exit(1);
 	}
