+++ tests/ts_print_raw.c
@@ -19,14 +19,9 @@ int main()
 {
 	struct tsdev *ts;
 
-	ts = ts_find(0);
+	ts = ts_setup(NULL, 0);
 	if (!ts) {
-		perror("ts_open");
-		exit(1);
-	}
-
-	if (ts_config(ts)) {
-		perror("ts_config");
+		perror("ts_setup");
 		exit(1);
 	}
